src/scripts/requirements.txt
Flask==3.0.3
google_api_python_client==2.147.0
google_auth_oauthlib==1.2.1
moviepy==1.0.3
numpy==1.22.0
openai_whisper==20231117
Pillow==10.4.0
praw==7.7.1
protobuf==5.28.2
torch==2.4.1
TTS==0.22.0
TTS==0.22.0
whisper==1.1.10
